完成了查询进度的接口

+++ exam/src/main/java/com/beyondhost/exam/entity/RoomTypeInfo.java
@@ -11,5 +11,5 @@ public class RoomTypeInfo {
     public long PartnerId;
     public long PoiId;
     public String RoomName;
-    public BigDecimal OriginalPrice;
+    public int OriginalPrice;
 }
